rename solver folder to core
