+++ src/core/modes/inter/index.test.ts
