+++ src/core/modes/inter/simulate.test.ts
