+++ src/core/modes/inter/simulate.ts
