+++ src/core/modes/inter/utils.test.ts
