+++ src/core/modes/inter/utils.ts
