+++ src/core/modes/intra/utils.test.ts
