+++ src/core/modes/intra/utils.ts
