+++ src/core/modes/rp/simulate.ts
