+++ src/core/process/log.ts
