+++ src/core/process/receipt.ts
